Convert example from Python to Java
This repo now uses a Java client SDK, rather than
the Python one that was being used before.  We only used Python before
this commit because we based this repo on [it's Python sibling][1].

This commit also shows that the code change to use a different language
in this workflow is really not very big.

[1] https://github.com/agilepathway/gauge-openapi-example