Remove validation proxy mode
Removing because there is a bug whereby the Java OpenAPI Generator sends
GET requests with a content-type, which causes Prism to reject these
requests.

The bug is the same (in Java) as the [Python bug][1] which was fixed
earlier in 2021.  It may well be more difficult to fix in Java, however,
as Java is the core language which OpenAPI Generator is written in.

Using the Prism validation proxy mode is more important than using the
OpenAPI Generator, so we may well want to refactor this repo to
reinstate the validation proxy and to remove the Generator.  This would
mean that we would hand-roll our own client code, but that's ok - as the
[Thoughtworks Tech Radar says][2], this type of bug is one of the risks
of using generated code rather than rolling our own.

[1]: https://github.com/OpenAPITools/openapi-generator/pull/9852
[2]: https://www.thoughtworks.com/radar/techniques/spec-based-codegen