Convert example from Python to Java (#1)
* Convert example from Python to Java

This repo now uses a Java client SDK, rather than
the Python one that was being used before.  We only used Python before
this commit because we based this repo on [it's Python sibling][1].

This commit also shows that the code change to use a different language
in this workflow is really not very big.

[1] https://github.com/agilepathway/gauge-openapi-example

* Remove validation proxy mode

Removing because there is a bug whereby the Java OpenAPI Generator sends
GET requests with a content-type, which causes Prism to reject these
requests.

The bug is the same (in Java) as the [Python bug][1] which was fixed
earlier in 2021.  It may well be more difficult to fix in Java, however,
as Java is the core language which OpenAPI Generator is written in.

Using the Prism validation proxy mode is more important than using the
OpenAPI Generator, so we may well want to refactor this repo to
reinstate the validation proxy and to remove the Generator.  This would
mean that we would hand-roll our own client code, but that's ok - as the
[Thoughtworks Tech Radar says][2], this type of bug is one of the risks
of using generated code rather than rolling our own.

[1]: https://github.com/OpenAPITools/openapi-generator/pull/9852
[2]: https://www.thoughtworks.com/radar/techniques/spec-based-codegen

* Add the OpenAPITools config file to source control

As per the guidance in the [OpenAPITools README][1]

[1]: https://github.com/OpenAPITools/openapi-generator-cli