Fix build by making pet name not required
Some of the pets currently in the production pet store currently don't
have names, which was correctly causing Prism to return an error saying
so when running in validation proxy mode.

This commit fixes that issue by making the name (and photo URL) not
required any more in our OpenAPI specification.