Categorize specs by the different consumers
This commit places our `available_pets` example spec into a directory
that shows clearly that the spec is part of the contract for a specific
web app. We would create a separate directory for each consumer. If some
of the consumers have duplicate needs then we'd simply duplicate those
specifications into each of their consumer directories, as each consumer
needs to have its own contract clearly and individually defined.