Run specs only in mock mode, not default or proxy
We want the contract CI to just run in mock mode as this allows the
consumer to do trunk-based development i.e. to merge a new consumer-
driven change to the OpenAPI contract into trunk on both the contract
repo and their consumer repo, without needing to wait for the provider
to have actually implemented the new feature.

So we no longer run the contract Gauge specs against the real provider
or in [Prism's validation proxy mode][1], as our latest evolution of the
consumer-driven workflow pushes this responsibility to the provider's
CI, rather than the contract CI.

[1]: https://meta.stoplight.io/docs/prism/docs/guides/03-validation-proxy.md