Use python not java (#11)
* Use Python devcontainer not Java

As part of the switch to use Python not Java this commit replaces the
Java [devcontainer][1] with the [default Python devcontainer][2], as the
VS Code IDE.

We will reinstate the installation of the software we need (e.g. Prism,
Gauge etc) in commits to follow this one.

[1]: https://code.visualstudio.com/docs/remote/containers
[2]: https://github.com/microsoft/vscode-dev-containers/tree/main/containers/python-3

* Reinstate Prism on devcontainer

This commit makes Prism available again for local development and
testing.

* Reinstate Gauge on devcontainer

Make Gauge available again in local development environment.

* Use Black for Python linting

[Black][1] is a great opinionated Python code formatter.

[1]: https://black.readthedocs.io/

* Use Python not Java for the tests

Decided to use Python as it is a less verbose, less heavyweight,
language than Java - which makes it a better choice for testing, all
things being equal.