Fix Gauge Python protobuf issue
The background to this is that Gauge uses [Protobuf][1] for internal
data exchange, under the hood.  Protobuf's [4.21.0 major release][2]
(released in May 2022) has some breaking changes, one of which causes
the latest version of [Gauge's Python plugin][3] to [not be able to
start][4].

This commit fixes the problem via a [temporary workaround][5] of setting
the `PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION` to be Python, as it was
before the breaking change.

[1]: https://github.com/protocolbuffers/protobuf
[2]: https://developers.google.com/protocol-buffers/docs/news/2022-05-06#python-updates
[3]: https://github.com/getgauge/gauge-python
[4]: https://github.com/getgauge/gauge-python/issues/256
[5]: https://github.com/protocolbuffers/protobuf/issues/10051#issuecomment-1138671075